requirements.txt
bitstring==3.1.9
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
docopt==0.6.2
docx2pdf==0.1.7
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
iso8601==0.1.16
itsdangerous==2.0.1
Jinja2==3.0.2
mutagen==1.45.1
Pillow==8.3.2
protobuf==3.17.3
pycryptodome==3.10.1
PyJWT==1.7.1
pyparsing==2.4.7
pysrt==1.1.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.0
pytube==11.0.1
SQLAlchemy==1.4.25
Werkzeug==2.0.2
WTForms==2.3.3
